+++ src/hooks/useModal.ts
@@ -1,0 +1,50 @@
+import { useEffect, useState } from 'react';
+
+const useModal = (isOpen: boolean, onClose: () => void) => {
+  const [mounted, setMounted] = useState(false);
+
+  useEffect(() => {
+    setMounted(true);
+    return () => setMounted(false);
+  }, []);
+
+  useEffect(() => {
+    const handleKeyDown = (e: KeyboardEvent) => {
+      if (e.key === 'Escape') onClose();
+    };
+
+    const handleClickOutside = (e: MouseEvent) => {
+      const modal = document.getElementById('modal-content');
+      if (modal && !modal.contains(e.target as Node)) {
+        onClose();
+      }
+    };
+
+    if (isOpen) {
+      window.addEventListener('keydown', handleKeyDown);
+      window.addEventListener('click', handleClickOutside);
+
+      const scrollBarWidth = window.innerWidth - document.body.clientWidth;
+      if (scrollBarWidth > 0) {
+        document.body.style.paddingRight = `${scrollBarWidth}px`;
+      }
+
+      document.body.style.overflow = 'hidden';
+    } else {
+      window.removeEventListener('keydown', handleKeyDown);
+      window.removeEventListener('click', handleClickOutside);
+      document.body.style.paddingRight = '0';
+      document.body.style.overflow = 'auto';
+    }
+
+    return () => {
+      window.removeEventListener('keydown', handleKeyDown);
+      window.removeEventListener('click', handleClickOutside);
+      document.body.style.paddingRight = '0';
+      document.body.style.overflow = 'auto';
+    };
+  }, [isOpen, onClose]);
+  return { mounted };
+};
+
+export default useModal;
